Extract Common Listing and Retrieval Functionality (#4220)
* Factor out common cloud storage client functionality

* Remove format_prefix

* Review feedback
